Rename components to containers
