+++ src/renderer/actions/ActionSaveReplayBuffer.tsx
@@ -3,11 +3,11 @@ import * as React from "react";
 import { ActionTypeGenerator, Context } from "@vinceau/event-actions";
 import { produce } from "immer";
 
-import { DelayInput, NotifyInput } from "@/components/Misc/InlineInputs";
+import { DelayInput, NotifyInput } from "@/containers/Misc/InlineInputs";
 import { delay as waitMillis, notify as sendNotification, parseSecondsDelayValue } from "@/lib/utils";
 import { ActionComponent } from "./types";
 
-import { CustomIcon } from "@/components/Misc/Misc";
+import { CustomIcon } from "@/containers/Misc/Misc";
 import { saveReplayBuffer } from "@/lib/obs";
 
 import obsIcon from "@/styles/images/obs.svg";
