+++ src/renderer/containers/Automator/ActionInputs.tsx
