+++ src/renderer/containers/Automator/Automator.tsx
