+++ src/renderer/containers/Automator/EventActions.tsx
