+++ src/renderer/containers/Misc/CharacterIcon.tsx
