+++ src/renderer/containers/Misc/InlineInputs.tsx
