+++ src/renderer/containers/Settings/ComboFinder.tsx
