+++ src/renderer/containers/Settings/ComboForm/CharacterSelect.tsx
