+++ src/renderer/containers/Settings/ComboForm/FormAdapters.tsx
