+++ src/renderer/containers/Settings/ComboForm/PerCharPercent.tsx
