+++ src/renderer/containers/Settings/ComboForm/PercentageSlider.tsx
