+++ src/renderer/containers/Settings/SlippiPage.tsx
